requirements.txt
groq>=0.4.0
langchain>=0.1.0
chromadb>=0.4.18
sentence-transformers>=2.2.2
bs4>=0.0.1
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.23.2
pydantic>=2.4.2
